# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A conversational Q\u0026A service for the Moto ride-hailing app.\nAsk Moto answers user questions about Moto's features, pricing modes, and functionality using a knowledge base retrieval system.",
        "title": "Ask Moto API",
        "contact": {
            "name": "Moto Team",
            "email": "support@moto.app"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/chat": {
            "post": {
                "description": "Submit a question and receive an AI-generated response based on the Moto knowledge base",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Send a message to Ask Moto",
                "parameters": [
                    {
                        "description": "Chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.ChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/http.ChatResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/feedback": {
            "post": {
                "description": "Submit user feedback (thumbs up/down) on a previous response",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Feedback"
                ],
                "summary": "Submit feedback on a response",
                "parameters": [
                    {
                        "description": "Feedback request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.FeedbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feedback submitted successfully",
                        "schema": {
                            "$ref": "#/definitions/http.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats": {
            "get": {
                "description": "Retrieve aggregated statistics about queries and feedback",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get analytics statistics",
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/domain.Stats"
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is running and healthy",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy",
                        "schema": {
                            "$ref": "#/definitions/http.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.AnswerType": {
            "type": "string",
            "enum": [
                "kb_answer",
                "action_guidance",
                "unknown"
            ],
            "x-enum-varnames": [
                "AnswerTypeKB",
                "AnswerTypeActionGuidance",
                "AnswerTypeUnknown"
            ]
        },
        "domain.Stats": {
            "type": "object",
            "properties": {
                "answer_rate": {
                    "type": "number"
                },
                "answered_queries": {
                    "type": "integer"
                },
                "negative_feedback": {
                    "type": "integer"
                },
                "positive_feedback": {
                    "type": "integer"
                },
                "total_feedback": {
                    "type": "integer"
                },
                "total_queries": {
                    "type": "integer"
                },
                "unknown_queries": {
                    "type": "integer"
                }
            }
        },
        "http.ChatRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "What is Moto?"
                },
                "user_id": {
                    "type": "string",
                    "example": "user-123"
                }
            }
        },
        "http.ChatResponse": {
            "type": "object",
            "properties": {
                "answer_type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.AnswerType"
                        }
                    ],
                    "example": "kb_answer"
                },
                "confidence": {
                    "type": "number",
                    "example": 0.95
                },
                "response": {
                    "type": "string",
                    "example": "Moto is a ride-hailing app that helps riders request rides, choose pricing modes, and helps drivers receive ride orders."
                },
                "response_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid request body"
                }
            }
        },
        "http.FeedbackRequest": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string",
                    "example": "Very helpful answer!"
                },
                "rating": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 2
                },
                "response_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "http.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "healthy"
                }
            }
        },
        "http.SuccessResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Conversational Q\u0026A endpoints",
            "name": "Chat"
        },
        {
            "description": "User feedback on responses",
            "name": "Feedback"
        },
        {
            "description": "Usage statistics and metrics",
            "name": "Analytics"
        },
        {
            "description": "Service health monitoring",
            "name": "Health"
        }
    ]
}